+++ src/components/index.js
@@ -1,0 +1,22 @@
+import HealthComp from '@/components/HealthComp'
+import DeckComp from '@/components/DeckComp'
+import HandComp from '@/components/HandComp'
+
+import FieldModal from '@/components/ModalWindows/FieldModal'
+import DeckGrave from '@/components/ModalWindows/DeckGrave'
+import CardModal from '@/components/ModalWindows/CardModal'
+
+import MenuBar from '@/pages/MenuBar'
+
+
+export default [
+    HealthComp,
+    DeckComp,
+    HandComp,
+    
+    FieldModal,
+    DeckGrave,
+    CardModal,
+
+    MenuBar,
+]
