+++ src/pages/AboutPage.vue
@@ -1,0 +1,15 @@
+<template>
+    <div>
+Страница о нас
+    </div>
+</template>
+
+<script>
+    export default {
+        
+    }
+</script>
+
+<style lang="scss" scoped>
+
+</style>
