+++ src/pages/MainPage.vue
@@ -1,0 +1,17 @@
+<template>
+    <div>
+ЭТО ГЛАВНАЯ СТРАНИЦА<br>
+<router-link to='/about'>О нас</router-link><br>
+<router-link to='/game'>Игра</router-link><br>
+    </div>
+</template>
+
+<script>
+    export default {
+        
+    }
+</script>
+
+<style>
+
+</style>
